First, somewhat working, version
This first version achieves a first good behavior which is to color code
using the right grammar in almost the right places, but the remaining
problem is that the "xml" (inside html! macros) is recognized as rust,
although it is tokenized using xml rules.

This is a result of me not really knowing what I am doing.